CI fix: Use version 20.04 when building with clang for Ubuntu
This fixes the problem that `clang-10` is not available for newer versions of
Ubuntu. Of course we could upgrade `clang` as well, but let's restore the
functionality as it was, and after that, we can introduce new changes.

By using a fixed version we get a more controlled build environment.